Use `PeekableByteStream` rather than `std.io.Reader`
Using a peekable byte stream allows us to peek at the current byte in
our stream without advancing it which should prove to be extremely
useful as we continue to add more and more checks for prefix bytes and
longer opcodes. Without using a peekable stream we would otherwise have
to resort to using an ugly temporary variable to store the current byte
until we move on to the next.